% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms.R
\name{cms}
\alias{cms}
\title{Add weights to the data, bound at a maximum.}
\usage{
cms(S, weight.upper.bound = 30)
}
\arguments{
\item{S}{}

\item{weight.upper.bound}{}
}
\value{
A matrix
}
\description{
Add weights to the data, bound at a maximum.
}
