% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conduit_2.R
\name{Conduit_2}
\alias{Conduit_2}
\title{Conduit between minimise_elements function and Fac_F_R
of steepest descent algorithm.}
\usage{
Conduit_2(F, place, S, cm)
}
\arguments{
\item{F}{}

\item{place}{}

\item{S}{}

\item{cm}{}
}
\description{
Conduit between minimise_elements function and Fac_F_R
of steepest descent algorithm.
}
