% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{convergence_figure}
\alias{convergence_figure}
\title{Convergence Figure}
\usage{
convergence_figure(fm_iter, niter = NULL)
}
\arguments{
\item{fm_iter}{A data.frame with columns of iter, phyto, pigment and ratio}

\item{niter}{Optional: the number of iterations on the x axis.
If \code{NULL}, will extract from the \code{iter} column of \code{fm_iter}.}
}
\value{
A figure with each pigment ratio per iteration per group
}
\description{
A figure to show the pigment ratios for each phytoplankton group for each
iteration.
}
\examples{
# ADD_EXAMPLES_HERE
}
