% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{F}
\alias{F}
\title{F data}
\format{
\subsection{\code{F}}{

A data frame with 10 rows and 15 columns:
\describe{
\item{chl.c1}{XX}
\item{Per}{XX}
\item{X19but}{XX}
...
}
}
}
\source{
XX
}
\usage{
F
}
\description{
F data
}
\keyword{datasets}
