% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fac_F_RR.R
\name{Fac_F_RR}
\alias{Fac_F_RR}
\title{Part of the steepest descent algorithm and work to reduce error given
the S and F matrices.}
\usage{
Fac_F_RR(Fmat, vary, S, cm, fac_rr = c(1, 2, 3), place = NULL)
}
\arguments{
\item{Fmat}{A list containing \verb{F matrix}, \code{RMSE} and \verb{C matrix}}

\item{vary}{Are the index of the non-zero integers to vary}

\item{S}{A matrix of samples (rows) and pigments (columns)}

\item{cm}{A vector of bounded weights for each pigment}

\item{fac_rr}{A numeric to call which scaler values to use}

\item{place}{A vector of all the indexs of non-zero pigment ratios}
}
\description{
Part of the steepest descent algorithm and work to reduce error given
the S and F matrices.
}
\keyword{internal}
