% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fac_F_RR3.R
\name{Fac_F_RR3}
\alias{Fac_F_RR3}
\title{Apply randomisation function to matrices and keep element that reduce error x4}
\usage{
Fac_F_RR3(F, vary, place, S, cm)
}
\arguments{
\item{vary}{}

\item{cm}{}
}
\description{
Apply randomisation function to matrices and keep element that reduce error x4
}
