% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fac_F_RR4.R
\name{Fac_F_RR4}
\alias{Fac_F_RR4}
\title{Apply randomisation function to matrices and keep element that reduce error x4}
\usage{
Fac_F_RR4(F, vary)
}
\arguments{
\item{vary}{}
}
\description{
Apply randomisation function to matrices and keep element that reduce error x4
}
