% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fm}
\alias{Fm}
\title{Fm data}
\format{
\subsection{\code{Fm}}{

A data frame with 9 rows and 15 columns:
\describe{
\item{chl.c1}{XX}
\item{Per}{XX}
\item{X19but}{XX}
...
}
}
}
\source{
XX
}
\usage{
Fm
}
\description{
Fm data
}
\keyword{datasets}
