% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Fp}
\alias{Fp}
\title{Fp data}
\format{
\subsection{\code{Fp}}{

A data frame with 9 rows and 15 columns:
\describe{
\item{chl_c1}{XX}
\item{Per}{XX}
\item{X19but}{XX}
...
}
}
}
\source{
XX
}
\usage{
Fp
}
\description{
Fp data
}
\keyword{datasets}
