% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_checks.R
\name{Homogenise_matrices}
\alias{Homogenise_matrices}
\title{Remove any column values that average 0. Further to this, also remove
phytoplankton groups from the F matrix if their diagnostic pigment
isn’t present.}
\usage{
Homogenise_matrices(S, F)
}
\arguments{
\item{F}{}
}
\description{
Remove any column values that average 0. Further to this, also remove
phytoplankton groups from the F matrix if their diagnostic pigment
isn’t present.
}
