% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_checks.R
\name{Matrix_checks}
\alias{Matrix_checks}
\title{This function ensures S and F matrices are properly formatted and ordered
for the simulated annealing function.}
\usage{
Matrix_checks(S, Fmat)
}
\arguments{
\item{S}{Sample data matrix – a matrix of pigment samples}

\item{Fmat}{Pigment to taxa matrix}
}
\value{
Named list with new S and Fmat matrices
}
\description{
Some checks applied:
\itemize{
\item drops columns with 0 values
\item drops taxa with missing major pigments, which are indicated with a '2'
\item drops pigments with < 1\% in samples
}
}
\examples{
MC <- Matrix_checks(Sm, Fm)  
Snew <- MC$Snew

}
