% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_max}
\alias{min_max}
\title{min and max data}
\format{
\subsection{\code{min_max}}{

A data frame with 68 rows and 4 columns:
\describe{
\item{Class}{XX}
\item{Pig_Abbrev}{XX}
\item{min}{XX}
\item{max}{XX}#'
...
}
}
}
\source{
XX
}
\usage{
min_max
}
\description{
min and max data
}
\keyword{datasets}
