% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minimise_elements.R
\name{Minimise_elements}
\alias{Minimise_elements}
\title{A function that reduces every for every element that didn't reduce in index function}
\usage{
Minimise_elements(F, place)
}
\description{
A function that reduces every for every element that didn't reduce in index function
}
