% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minimise_elements1.R
\name{Minimise_elements1}
\alias{Minimise_elements1}
\title{A function that reduces every for every element that didn't reduce in index function}
\usage{
Minimise_elements1(F)
}
\description{
A function that reduces every for every element that didn't reduce in index function
}
