% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minimise_elements2.R
\name{Minimise_elements2}
\alias{Minimise_elements2}
\title{A function that reduces every for every element that didn't reduce in index function}
\usage{
Minimise_elements2(F, place, S, cm)
}
\arguments{
\item{F}{}

\item{place}{}

\item{cm}{}
}
\description{
A function that reduces every for every element that didn't reduce in index function
}
