% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minimise_elements4.R
\name{Minimise_elements4}
\alias{Minimise_elements4}
\title{A function that reduces every for every element that didn't reduce in index function}
\usage{
Minimise_elements4(F)
}
\description{
A function that reduces every for every element that didn't reduce in index function
}
