% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minimise_elements_comb.R
\name{Minimise_elements_comb}
\alias{Minimise_elements_comb}
\title{Part of the steepest descent algorithm}
\usage{
Minimise_elements_comb(Fmat, place, S, cm, c1_num = c(1, 2, 3))
}
\arguments{
\item{Fmat}{xx}

\item{place}{xx}

\item{S}{xx}

\item{cm}{xx}
}
\description{
Part of the steepest descent algorithm
}
\keyword{internal}
