% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNLS_MF.R
\name{NNLS_MF}
\alias{NNLS_MF}
\title{Performs the non-negative matrix factorisation for given phytoplankton
pigments and pigment ratios, to attain an estimate of phytoplankton
class abundances.}
\usage{
NNLS_MF(Fn, S, cm)
}
\arguments{
\item{Fn}{}

\item{S}{}

\item{cm}{}
}
\value{
A list
}
\description{
Performs the non-negative matrix factorisation for given phytoplankton
pigments and pigment ratios, to attain an estimate of phytoplankton
class abundances.
}
