% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{Normalise_S}
\alias{Normalise_S}
\title{Normalise matrix to row sum}
\usage{
Normalise_S(S)
}
\arguments{
\item{S}{A matrix or data.frame to be normalized}
}
\value{
A matrix
}
\description{
This function normalises each column in S to row sum
}
\examples{
# Create a sample matrix
S <- as.matrix(phytoclass::Sm)
normalized <- Normalise_S(S)

}
\keyword{internal}
