% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{phyto_figure}
\alias{phyto_figure}
\title{Phytoplankton Class Abundance Figure}
\usage{
phyto_figure(c_matrix)
}
\arguments{
\item{c_matrix}{C matrix of class abundance concentrations}
}
\value{
A stacked line plot with sample number on x axis, chl a
concentrations on y axis, and phytoplankton groups as colors
}
\description{
This function plots the class abundances as output by \code{simulated_annealing}.
}
\examples{
# ADD_EXAMPLES_HERE
}
