% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prochloro_Random_Neighbour2.R
\name{Prochloro_Random_Neighbour_2}
\alias{Prochloro_Random_Neighbour_2}
\title{Selects a random neighbours for all non-zero pigments in the simulated
annealing algorithm.}
\usage{
Prochloro_Random_Neighbour_2(
  f_new,
  Temp,
  chlv,
  chlvp,
  N,
  place,
  S,
  S_weights,
  minF,
  maxF
)
}
\arguments{
\item{f_new}{F matrix}

\item{Temp}{Temperature of the annealing}

\item{chlv}{Chlorophyll-a column}

\item{chlvp}{Dvchla column}

\item{N}{Indexs of pigment ratios to be changed}

\item{place}{Indexes in F matrix where values are non-zero}

\item{S}{S matrix of samples}

\item{S_weights}{Weights for NNLS algorithm}

\item{minF}{Minimum bounds for each phytoplankton group and pigments}

\item{maxF}{Maximum bounds for each phytoplankton group and pigments}
}
\description{
Selects a random neighbours for all non-zero pigments in the simulated
annealing algorithm.
}
\keyword{internal}
