% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand.R
\name{rand}
\alias{rand}
\title{Randomisation function}
\usage{
rand(x, min.scaler, max.scaler)
}
\arguments{
\item{x}{}

\item{max.scaler}{}

\item{na.rm}{}
}
\value{
numeric
}
\description{
Randomisation function
}
