% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_Rand1.R
\name{Replace_Rand1}
\alias{Replace_Rand1}
\title{Apply randomization functions to matrices}
\usage{
Replace_Rand1(F, i, S, cm)
}
\arguments{
\item{i}{}

\item{cm}{}
}
\description{
Apply randomization functions to matrices
}
