% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_Rand2.R
\name{Replace_Rand2}
\alias{Replace_Rand2}
\title{Apply randomization functions to matrices}
\usage{
Replace_Rand2(F, i)
}
\arguments{
\item{i}{}
}
\description{
Apply randomization functions to matrices
}
