% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_Rand3.R
\name{Replace_Rand3}
\alias{Replace_Rand3}
\title{Apply randomization functions to matrices}
\usage{
Replace_Rand3(F, i, S, cm)
}
\arguments{
\item{i}{}

\item{S}{}

\item{cm}{}
}
\description{
Apply randomization functions to matrices
}
