% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_Rand4.R
\name{Replace_Rand4}
\alias{Replace_Rand4}
\title{Apply randomization functions to matrices}
\usage{
Replace_Rand4(F, i, S, cm)
}
\arguments{
\item{i}{}
}
\description{
Apply randomization functions to matrices
}
