% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{S}
\alias{S}
\title{S data}
\format{
\subsection{\code{S}}{

A data frame with 17 rows and 16 columns:
\describe{
\item{chl.c1}{XX}
\item{Per}{XX}
\item{X19but}{XX}
...
}
}
}
\source{
XX
}
\usage{
S
}
\description{
S data
}
\keyword{datasets}
