% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAALS.R
\name{SAALS}
\alias{SAALS}
\title{Perform the simulated annealing algorithm for given
F and S matrices.}
\usage{
SAALS(Ft, min.value, max.value, place, S, cm, num.loops)
}
\arguments{
\item{Ft}{}

\item{S}{}

\item{num.loops}{}
}
\description{
Perform the simulated annealing algorithm for given
F and S matrices.
}
