% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_annealing.R
\name{simulated_annealing}
\alias{simulated_annealing}
\title{This is the main phytoclass algorithm. It performs simulated annealing
algorithm for S and F matrices. See the examples (Fm, Sm) for how to set
up matrices, and the vignette for more detailed instructions. Different
pigments and phytoplankton groups may be used.}
\usage{
simulated_annealing(
  S,
  Fmat = NULL,
  user_defined_min_max = NULL,
  do_matrix_checks = TRUE,
  niter = 500,
  step = 0.009,
  weight.upper.bound = 30,
  verbose = TRUE,
  seed = NULL,
  check_converge = 100,
  alt_pro_name = NULL,
  preserve_F = TRUE,
  chemtax_style = TRUE
)
}
\arguments{
\item{S}{Sample data matrix – a matrix of pigment samples}

\item{Fmat}{Pigment to Chl a matrix}

\item{user_defined_min_max}{data frame with some format as min_max built-in data}

\item{do_matrix_checks}{This should only be set to TRUE when using the default
values. This will remove pigment columns that have
column sums of 0. Set to FALSE if using customised
names for pigments and phytoplankton groups}

\item{niter}{Number of iterations (default is 500)}

\item{step}{Step ratio used (default is 0.009)}

\item{weight.upper.bound}{Upper limit of the weights applied (default value is 30).}

\item{verbose}{Logical value. Output error and temperature at each
iteration. Default value of TRUE}

\item{seed}{Set number to reproduce the same results}

\item{check_converge}{TRUE/FALSE/integer; set the number of F matrices to
for convergence checking}

\item{alt_pro_name}{Optional: additional alternate versions of
divinyl-chlorophyll-a spellings used to detect
prochlorococcus (Default: "dvchl", "dvchla", "dv_chla")}

\item{preserve_F}{Initial nnls C matrix from from non-binary F matrix values
clamped to user_defined_min_max instead of starting with binary 0/1
(default = TRUE)}

\item{chemtax_style}{uncapped column-mean weighting with no chl_a
override (default = TRUE)}
}
\value{
A list containing
\enumerate{
\item Fmat matrix
\item RMSE (Root Mean Square Error)
\item condition number
\item Class abundances
\item Figure (plot of results)
\item MAE (Mean Absolute Error)
\item Error
\item F_mat_iter
\item converge_plot
}
}
\description{
This is the main phytoclass algorithm. It performs simulated annealing
algorithm for S and F matrices. See the examples (Fm, Sm) for how to set
up matrices, and the vignette for more detailed instructions. Different
pigments and phytoplankton groups may be used.
}
\examples{
# Using the built-in matrices Sm and Fm
set.seed(5326)
sa.example <- simulated_annealing(Sm, Fm, niter = 5)
sa.example$Figure


}
