% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sm}
\alias{Sm}
\title{Sm data}
\format{
\subsection{\code{Sm}}{

A data frame with 29 rows and 15 columns:
\describe{
\item{chl.c1}{XX}
\item{Per}{XX}
\item{X19but}{XX}
...
}
}
}
\source{
XX
}
\usage{
Sm
}
\description{
Sm data
}
\keyword{datasets}
