% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sp}
\alias{Sp}
\title{Sp data}
\format{
\subsection{\code{Sp}}{

A data frame with 29 rows and 15 columns:
\describe{
\item{chl_c1}{XX}
\item{Per}{XX}
\item{X19but}{XX}
...
}
}
}
\source{
XX
}
\usage{
Sp
}
\description{
Sp data
}
\keyword{datasets}
