% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steepest_Desc.R
\name{Steepest_Desc}
\alias{Steepest_Desc}
\title{Stand-alone version of steepest descent algorithm}
\usage{
Steepest_Desc(F, S, num.loops)
}
\arguments{
\item{S}{}

\item{num.loops}{}
}
\description{
Stand-alone version of steepest descent algorithm
}
