% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test1.R
\name{Test1}
\alias{Test1}
\title{conduit between minimise_elements function and Fac_F_R}
\usage{
Test1(F, place, S, cm)
}
\arguments{
\item{F}{}
}
\description{
conduit between minimise_elements function and Fac_F_R
}
