% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Try_This.R
\name{Try_This}
\alias{Try_This}
\title{sets how many iterations the algorithm should continue for x4.
This is the 'steepest descent' algorithm}
\usage{
Try_This(F)
}
\description{
sets how many iterations the algorithm should continue for x4.
This is the 'steepest descent' algorithm
}
