% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Try_This2.R
\name{Try_This2}
\alias{Try_This2}
\title{sets how many iterations the algorithm should continue for x4.
This is the 'steepest descent' algorithm}
\usage{
Try_This2(F, place, S, cm)
}
\arguments{
\item{F}{}

\item{place}{}

\item{S}{}

\item{cm}{}
}
\description{
sets how many iterations the algorithm should continue for x4.
This is the 'steepest descent' algorithm
}
