% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorise.R
\name{vectorise}
\alias{vectorise}
\title{This returns all elements that aren't 0 in the F matrix}
\usage{
vectorise(Fg)
}
\arguments{
\item{Fg}{}
}
\description{
This returns all elements that aren't 0 in the F matrix
}
