% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrangle_S.R
\name{Wrangle_S}
\alias{Wrangle_S}
\title{This function normalises each column in S to row sum}
\usage{
Wrangle_S(S)
}
\arguments{
\item{S}{}
}
\value{
A matrix
}
\description{
This function normalises each column in S to row sum
}
