% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bounded_weights.R
\name{Bounded_weights}
\alias{Bounded_weights}
\title{Add weights to the data, bound at a maximum.}
\usage{
Bounded_weights(S, weight.upper.bound)
}
\arguments{
\item{S}{XX}

\item{weight.upper.bound}{XX}
}
\value{
A matrix
}
\description{
Add weights to the data, bound at a maximum.
}
