% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Condition_test.R
\name{Condition_test}
\alias{Condition_test}
\title{Calculate the condition number ...}
\usage{
Condition_test(Fn, min.val, max.val)
}
\arguments{
\item{max.val}{}
}
\description{
Calculate the condition number ...
}
