% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Condition_test.R
\name{Condition_test}
\alias{Condition_test}
\title{Calculate the condition number ...}
\usage{
Condition_test(S, Fn, min.val = NULL, max.val = NULL)
}
\arguments{
\item{S}{XX}

\item{Fn}{XX}

\item{min.val}{XX}

\item{max.val}{XX}
}
\description{
Calculate the condition number ...
}
\keyword{internal}
