% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conduit.R
\name{Conduit}
\alias{Conduit}
\title{Conduit between minimise_elements function and Fac_F_R
of steepest descent algorithm.}
\usage{
Conduit(Fmat, place, S, cm, c_num = c(1, 2, 3))
}
\arguments{
\item{Fmat}{xx}

\item{place}{xx}

\item{S}{xx}

\item{cm}{xx}
}
\description{
Conduit between minimise_elements function and Fac_F_R
of steepest descent algorithm.
}
\keyword{internal}
