% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Default_min_max.R
\name{Default_min_max}
\alias{Default_min_max}
\title{Sets the default minimum and maximum values for phytoplankton groups
pigment ratios. To use this function, pigment and phytoplankton group
names will need to fit the naming criteria of phytoclass.}
\usage{
Default_min_max(min_max, Fmat, place)
}
\arguments{
\item{min_max}{xx}

\item{Fmat}{xx}

\item{place}{xx}
}
\description{
Sets the default minimum and maximum values for phytoplankton groups
pigment ratios. To use this function, pigment and phytoplankton group
names will need to fit the naming criteria of phytoclass.
}
\keyword{internal}
