% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fac_F_Final.R
\name{Fac_F_Final}
\alias{Fac_F_Final}
\title{This is the final fit, and multiplies the final value by Chl concentration so
they're in units of Chl a biomass}
\usage{
Fac_F_Final(Fn, S, S_Chl, cm)
}
\arguments{
\item{Fn}{}

\item{cm}{}
}
\description{
This is the final fit, and multiplies the final value by Chl concentration so
they're in units of Chl a biomass
}
