% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fac_F_RR1.R
\name{Fac_F_RR1}
\alias{Fac_F_RR1}
\title{Apply randomisation function to matrices and keep element that reduce error x4}
\usage{
Fac_F_RR1(F, vary, S, cm)
}
\arguments{
\item{F}{XX}

\item{vary}{XX}

\item{S}{XX}

\item{cm}{XX}
}
\description{
Apply randomisation function to matrices and keep element that reduce error x4
}
