% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fac_F_RR1.R
\name{Fac_F_RR1}
\alias{Fac_F_RR1}
\title{Part of the steepest descent algorithm and work to reduce error given
the S and F matrices.}
\usage{
Fac_F_RR1(Fmat, vary, S, cm)
}
\arguments{
\item{Fmat}{XX}

\item{vary}{XX}

\item{S}{XX}

\item{cm}{XX}
}
\description{
Part of the steepest descent algorithm and work to reduce error given
the S and F matrices.
}
\keyword{internal}
