% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fac_F_RR2.R
\name{Fac_F_RR2}
\alias{Fac_F_RR2}
\title{Part of the steepest descent algorithm and work to reduce error given
the S and F matrices}
\usage{
Fac_F_RR2(Fmat, vary, place, S, cm)
}
\arguments{
\item{Fmat}{xx}

\item{vary}{xx}

\item{place}{xx}

\item{S}{xx}

\item{cm}{xx}
}
\description{
Part of the steepest descent algorithm and work to reduce error given
the S and F matrices
}
\keyword{internal}
