% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrix_checks.R
\name{Matrix_checks}
\alias{Matrix_checks}
\title{Remove any column values that average 0. Further to this, also remove
phytoplankton groups from the F matrix if their diagnostic pigment
isn’t present.}
\usage{
Matrix_checks(S, Fmat)
}
\arguments{
\item{S}{Sample data matrix – a matrix of pigment samples}

\item{Fmat}{Pigment to Chl a matrix}
}
\value{
Named list with new S and Fmat matrices
}
\description{
Remove any column values that average 0. Further to this, also remove
phytoplankton groups from the F matrix if their diagnostic pigment
isn’t present.
}
\examples{
MC <- Matrix_checks(Sm, Fm)  
Snew <- MC$Snew

}
