% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minimise_elements.R
\name{Minimise_elements}
\alias{Minimise_elements}
\title{Part of the steepest descent algorithm}
\usage{
Minimise_elements(Fmat, place, S, cm)
}
\arguments{
\item{Fmat}{xx}

\item{place}{xx}

\item{S}{xx}

\item{cm}{xx}
}
\description{
Part of the steepest descent algorithm
}
\keyword{internal}
