% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Minimise_elements2.R
\name{Minimise_elements2}
\alias{Minimise_elements2}
\title{A function that reduces every for every element that didn't reduce in index function}
\usage{
Minimise_elements2(Fmat, place, S, cm)
}
\arguments{
\item{Fmat}{xx}

\item{place}{xx}

\item{S}{xx}

\item{cm}{xx}
}
\description{
A function that reduces every for every element that didn't reduce in index function
}
\keyword{internal}
