% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNLS_MF.R
\name{NNLS_MF}
\alias{NNLS_MF}
\title{Performs the non-negative matrix factorisation for given phytoplankton
pigments and pigment ratios, to attain an estimate of phytoplankton
class abundances.}
\usage{
NNLS_MF(Fn, S, cm = NULL)
}
\arguments{
\item{S}{Sample data matrix – a matrix of pigment samples}

\item{cm}{Weights for each column}

\item{F}{Pigment to Chl a matrix}
}
\value{
A list containing
\enumerate{
\item F matrix
\item RMSE (Root Mean Square Error)
\item C matrix for class abundances
}
}
\description{
Performs the non-negative matrix factorisation for given phytoplankton
pigments and pigment ratios, to attain an estimate of phytoplankton
class abundances.
}
\examples{
MC <- Matrix_checks(Sm,Fm)
Snew <- MC$Snew
Fnew <- MC$Fnew
cm <- Bounded_weights(Snew, weight.upper.bound = 30)
NNLS_MF(Fnew, Snew, cm)

}
