% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNLS_ML_Final.R
\name{NNLS_MF_Final}
\alias{NNLS_MF_Final}
\title{Performs the non-negative matrix factorisation for given phytoplankton
pigments and pigment ratios, to attain an estimate of phytoplankton
class abundances.}
\usage{
NNLS_MF_Final(Fn, S, S_Chl, cm)
}
\arguments{
\item{Fn}{}

\item{cm}{}
}
\description{
Unlike NNLS_ML(), it also removes any weighting and normalisation, and
also multiplies relative abundances by chlorophyll values to determine
the biomass of phytoplankton groups.
removes any weighting and normalisation, it also multiplies relative abundances by chlorophyll values to determine the biomass of phytoplankton groups.
}
