% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalise_F.R
\name{Normalise_F}
\alias{Normalise_F}
\title{This function normalises each column in F to row sum}
\usage{
Normalise_F(Fmat)
}
\arguments{
\item{Fmat}{xx}
}
\value{
A matrix
}
\description{
This function normalises each column in F to row sum
}
\keyword{internal}
