% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalise_S.R
\name{Normalise_S}
\alias{Normalise_S}
\title{This function normalises each column in S to row sum}
\usage{
Normalise_S(S)
}
\arguments{
\item{S}{xx}
}
\value{
A matrix
}
\description{
This function normalises each column in S to row sum
}
\keyword{internal}
