% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prochloro_NNLS_MF_Final.R
\name{Prochloro_NNLS_MF_Final}
\alias{Prochloro_NNLS_MF_Final}
\title{Final step for MF with prochlorococcus}
\usage{
Prochloro_NNLS_MF_Final(Fn, S, S_Chl, cm, S_dvChl)
}
\arguments{
\item{Fn}{}

\item{S}{}

\item{S_Chl}{}

\item{cm}{}
}
\description{
Final step for MF with prochlorococcus
}
\keyword{internal}
