% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prochloro_Normalise_F.R
\name{Prochloro_Normalise_F}
\alias{Prochloro_Normalise_F}
\title{Normalise F for prochloro}
\usage{
Prochloro_Normalise_F(Fmat)
}
\arguments{
\item{Fmat}{}
}
\description{
Normalise F for prochloro
}
\keyword{internal}
