% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prochloro_Random_Neighbour2.R
\name{Prochloro_Random_Neighbour_2}
\alias{Prochloro_Random_Neighbour_2}
\title{Selects a random neighbour for the simulated annealing algorithm.}
\usage{
Prochloro_Random_Neighbour_2(
  Fn,
  Temp,
  chlv,
  S,
  S_weights,
  minF_vec,
  maxF_vec,
  chlvp,
  Fi_mask
)
}
\arguments{
\item{Fn}{F matrix}

\item{Temp}{temperature of the annealing}

\item{chlv}{Chlorophyll-a column}

\item{S}{xx}

\item{S_weights}{xx}

\item{minF_vec}{xx}

\item{maxF_vec}{xx}

\item{chlvp}{xx}

\item{Fi_mask}{xx}
}
\description{
Selects a random neighbour for the simulated annealing algorithm.
}
\keyword{internal}
