% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{Prochloro_Wrangling}
\alias{Prochloro_Wrangling}
\title{Prochloro Wrangling}
\usage{
Prochloro_Wrangling(Fl, min.val, max.val)
}
\arguments{
\item{Fl}{Initial F matrix (i.e. pigment ratio matrix)}

\item{min.val}{A vector of minimum values for each non-zero pigment ratio}

\item{max.val}{A vector of maximum values for each non-zero pigment ratio}
}
\value{
A list containing vectorized min/max bounds and current values for
Prochlorococcus-aware normalization
}
\description{
Prochloro Wrangling
}
\keyword{internal}
