% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Randomise_elements.R
\name{Randomise_elements}
\alias{Randomise_elements}
\title{Randomise individual elements in the F matrix.}
\usage{
Randomise_elements(x, min.scaler, max.scaler)
}
\arguments{
\item{x}{}

\item{max.scaler}{}

\item{na.rm}{}
}
\value{
numeric
}
\description{
Randomise individual elements in the F matrix.
}
