% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_Rand.R
\name{Replace_Rand}
\alias{Replace_Rand}
\title{Apply randomization functions to matrices}
\usage{
Replace_Rand(F, i, S, cm, min.scaler, max.scaler)
}
\arguments{
\item{i}{}

\item{max.scaler}{}
}
\description{
Apply randomization functions to matrices
}
