% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_Rand.R
\name{Replace_Rand}
\alias{Replace_Rand}
\title{Select the new F matrix element with lowest error in the steepest
descent algortithm.}
\usage{
Replace_Rand(F, i, S, cm, min.scaler, max.scaler)
}
\arguments{
\item{i}{}

\item{max.scaler}{}
}
\description{
Select the new F matrix element with lowest error in the steepest
descent algortithm.
}
