% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_Rand.R
\name{Replace_Rand}
\alias{Replace_Rand}
\title{Select the new F matrix element with lowest error in the steepest
descent algorithm.}
\usage{
Replace_Rand(Fmat, i, S, cm, min.scaler, max.scaler)
}
\arguments{
\item{Fmat}{xx}

\item{i}{xx}

\item{S}{xx}

\item{cm}{xx}

\item{min.scaler}{xx}

\item{max.scaler}{xx}
}
\description{
Select the new F matrix element with lowest error in the steepest
descent algorithm.
}
\keyword{internal}
