% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAALS.R
\name{SAALS}
\alias{SAALS}
\title{Apply the steepest descent algorithm}
\usage{
SAALS(Ft, min.value, max.value, place, S, cm, num.loops)
}
\arguments{
\item{Ft}{}

\item{min.value}{}

\item{max.value}{}

\item{place}{}

\item{S}{}

\item{cm}{}

\item{num.loops}{}
}
\description{
Apply the steepest descent algorithm
}
