% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAALS.R
\name{SAALS}
\alias{SAALS}
\title{to apply to SD algorithm after each iteration of simulated annealing.}
\usage{
SAALS(Ft, min.value, max.value, place, S, cm)
}
\arguments{
\item{Ft}{}

\item{S}{}

\item{Cm}{}
}
\description{
to apply to SD algorithm after each iteration of simulated annealing.
}
