% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAALS.R
\name{SAALS}
\alias{SAALS}
\title{Apply the steepest descent algorithm}
\usage{
SAALS(Ft, min.value, max.value, place, S, cm, num.loops)
}
\arguments{
\item{Ft}{xx}

\item{min.value}{xx}

\item{max.value}{xx}

\item{place}{xx}

\item{S}{xx}

\item{cm}{xx}

\item{num.loops}{xx}
}
\description{
Apply the steepest descent algorithm
}
\keyword{internal}
