% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAALS2.R
\name{SAALS2}
\alias{SAALS2}
\title{to apply to SD algorithm after each iteration of simulated annealing.}
\usage{
SAALS2(Ft, min.value, max.value, place, S, cm)
}
\arguments{
\item{Ft}{}

\item{cm}{}
}
\description{
to apply to SD algorithm after each iteration of simulated annealing.
}
