% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_annealing.R
\name{simulated_annealing}
\alias{simulated_annealing}
\title{Perform simulated annealing algorithm for given F and S matrices}
\usage{
simulated_annealing(S, F, user_defined_min_max = NA, niter, step)
}
\arguments{
\item{user_defined_min_max}{data frame with some format as min_max built-in data}

\item{niter}{}

\item{step}{}
}
\description{
Perform simulated annealing algorithm for given F and S matrices
}
