% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_annealing.R
\name{simulated_annealing}
\alias{simulated_annealing}
\title{Perform simulated annealing algorithm for given F and S matrices}
\usage{
simulated_annealing(S, F, min_max, niter, step)
}
\arguments{
\item{niter}{}

\item{step}{}
}
\description{
Perform simulated annealing algorithm for given F and S matrices
}
