% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_annealing.R
\name{simulated_annealing}
\alias{simulated_annealing}
\title{input = F matrix, number of iterations, and the step to use.}
\usage{
simulated_annealing(S, F, min_max, niter, step)
}
\arguments{
\item{niter}{}

\item{step}{}
}
\description{
input = F matrix, number of iterations, and the step to use.
}
