% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_annealing.R
\name{simulated_annealing}
\alias{simulated_annealing}
\title{Perform simulated annealing algorithm for S and F matrices}
\usage{
simulated_annealing(
  S,
  F = NULL,
  user_defined_min_max = NULL,
  do_matrix_checks = TRUE,
  niter = 500,
  step = 0.009
)
}
\arguments{
\item{S}{xx}

\item{F}{xx}

\item{user_defined_min_max}{data frame with some format as min_max built-in data}

\item{do_matrix_checks}{xx}

\item{niter}{xx}

\item{step}{xx}
}
\value{
A list containing
\enumerate{
\item F matrix
\item RMSE
\item sMAPE
\item condition number
\item Class abudances
\item MAE
\item Error
}
}
\description{
Perform simulated annealing algorithm for S and F matrices
}
\examples{
# Using the built-in matrices Sm and Fm
set.seed(5326)
sa.example <- simulated_annealing(Sm, Fm)
sa.example$Figure
}
