% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_annealing2.R
\name{simulated_annealing2}
\alias{simulated_annealing2}
\title{input = F matrix, number of iterations, and the step to use.}
\usage{
simulated_annealing2(Fi, niter, step)
}
\arguments{
\item{Fi}{}

\item{niter}{}

\item{step}{}
}
\description{
input = F matrix, number of iterations, and the step to use.
}
