% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steepest_Desc.R
\name{Steepest_Desc}
\alias{Steepest_Desc}
\title{Stand-alone version of steepest descent algorithm}
\usage{
Steepest_Desc(F, S, num.loops)
}
\arguments{
\item{F}{Pigment to Chl a matrix}

\item{S}{Sample data matrix – a matrix of pigment samples}

\item{num.loops}{Number of loops/iterations to perform (no default)}
}
\value{
XX
}
\description{
Stand-alone version of steepest descent algorithm
}
\examples{
MC <- Matrix_checks(Sm,Fm)
Snew <- MC$Snew
Fnew <- MC$Fnew
SDRes <- Steepest_Desc(Fnew,Snew, num.loops = 20)

}
