% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steepest_Descent.R
\name{Steepest_Descent}
\alias{Steepest_Descent}
\title{Performs the steepest descent algorithm for a set number of iterations}
\usage{
Steepest_Descent(F, place, S, cm, num.loops)
}
\arguments{
\item{F}{}

\item{place}{}

\item{S}{}

\item{cm}{}

\item{num.loops}{}
}
\description{
Performs the steepest descent algorithm for a set number of iterations
}
