% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Steepest_Descent.R
\name{Steepest_Descent}
\alias{Steepest_Descent}
\title{sets how many iterations the algorithm should continue for x4.
This is the 'steepest descent' algorithm}
\usage{
Steepest_Descent(F, place, S, cm, num.loops)
}
\arguments{
\item{F}{}

\item{place}{}

\item{S}{}

\item{num.loops}{}
}
\description{
sets how many iterations the algorithm should continue for x4.
This is the 'steepest descent' algorithm
}
