% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorise.R
\name{vectorise}
\alias{vectorise}
\title{Vectorise a matrix and keep non-zero elements}
\usage{
vectorise(Fmat)
}
\arguments{
\item{Fmat}{A matrix to vectorise}
}
\value{
A vector of non-zero pigment elements
}
\description{
Turn each non-zero element of the F-matrix into a vector
}
\keyword{internal}
