% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorise.R
\name{vectorise}
\alias{vectorise}
\title{Turn each non-zero element of the F-matrix into a vector}
\usage{
vectorise(F)
}
\arguments{
\item{F}{}
}
\description{
Turn each non-zero element of the F-matrix into a vector
}
