% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weight_error.R
\name{Weight_error}
\alias{Weight_error}
\title{Apply weights to F/S matrices}
\usage{
Weight_error(S, cm)
}
\arguments{
\item{S}{xx}

\item{cm}{xx}
}
\value{
A matrix
}
\description{
Apply weights to F/S matrices
}
\keyword{internal}
