% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrangling.R
\name{Wrangling}
\alias{Wrangling}
\title{wrangles the data for the simulated annealing algorithm (maxs, mins etc...)}
\usage{
Wrangling(Fl, min.val, max.val)
}
\arguments{
\item{Fl}{}

\item{min.val}{}

\item{max.val}{}
}
\description{
wrangles the data for the simulated annealing algorithm (maxs, mins etc...)
}
