% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrangling.R
\name{Wrangling}
\alias{Wrangling}
\title{wrangles the data for the simulated annealing algorithm (maxs, mins etc...)}
\usage{
Wrangling(Fl)
}
\arguments{
\item{Fl}{}
}
\description{
wrangles the data for the simulated annealing algorithm (maxs, mins etc...)
}
