% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrangling.R
\name{Wrangling}
\alias{Wrangling}
\title{Converts data-types and selects data for randomisation in
the simulated annealing algorithm}
\usage{
Wrangling(Fl, min.val, max.val)
}
\arguments{
\item{Fl}{}

\item{min.val}{}

\item{max.val}{}
}
\description{
Converts data-types and selects data for randomisation in
the simulated annealing algorithm
}
